Revert "Adds event_type to CSV import as optional column"
This reverts commit c06074a4220de892d59f85c01ddc928912b78b98.
